Change style to match figma
