+++ web/src/features/map-controls/MapButton.tsx
@@ -1,5 +1,4 @@
-import * as Toggle from '@radix-ui/react-toggle';
-import { twMerge } from 'tailwind-merge';
+import { Button } from 'components/Button';
 
 import TooltipWrapper from '../../components/tooltips/TooltipWrapper';
 
@@ -9,35 +8,29 @@ interface MapButtonProperties {
   tooltipText?: string;
   className?: string;
   dataTestId?: string;
-  asToggle?: boolean;
   ariaLabel?: string;
+  backgroundClasses?: string;
 }
 
 export default function MapButton({
   icon,
   tooltipText,
-  className,
   dataTestId,
   onClick,
-  asToggle,
   ariaLabel,
+  backgroundClasses,
 }: MapButtonProperties) {
-  const Component = asToggle ? Toggle.Root : 'div';
   return (
     <TooltipWrapper tooltipContent={tooltipText}>
-      <Component
+      <Button
+        size="md"
+        type="transparent"
+        icon={icon}
         onClick={onClick}
-        className={twMerge(
-          `flex h-8 w-8 items-center justify-center rounded bg-white/80 text-left shadow-lg backdrop-blur-sm transition hover:bg-white dark:bg-gray-800/80 dark:hover:bg-gray-900/90`,
-          className,
-          asToggle && 'pointer-events-auto'
-        )}
         aria-label={ariaLabel}
         data-test-id={dataTestId}
-        role="button"
-      >
-        <div>{icon}</div>
-      </Component>
+        backgroundClasses={backgroundClasses}
+      />
     </TooltipWrapper>
   );
 }
